+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Both client libraries are pre-1.0, and they have separate versioning.
   - `Image.fromAwsEcr` in typescript and `NewImageFromAwsEcr` in golang.
   - `Image.fromGcpArtifactRegistry` in typescript and `NewImageFromGcpArtifactRegistry` in golang.
 - Added support for `Function.getCurrentStats()` (JS) / `Function.GetCurrentStats()` (Go).
+- Added support for `Function.updateAutoscaler()` (JS) / `Function.UpdateAutoscaler()` (Go).
 
 ## modal-js/v0.3.16, modal-go/v0.0.16